cli: move AWS chaining mode setup to AWS-specific code
We move the logic for setting up chaining mode on AWS from the high
level entrypoint to the dedicated AWS-specific code. This will help keep
the entrypoint clean as we intend to enrich the high level logic for AWS
in subsequent commits.
