+++ cilium-cli/install/install.go
@@ -18,7 +18,6 @@ import (
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/yaml"
 
@@ -185,11 +184,6 @@ func (k *K8sInstaller) listVersions() error {
 	return err
 }
 
-func getChainingMode(values map[string]interface{}) string {
-	chainingMode, _, _ := unstructured.NestedString(values, "cni", "chainingMode")
-	return chainingMode
-}
-
 func (k *K8sInstaller) preinstall(ctx context.Context) error {
 	// TODO (ajs): Note that we have our own implementation of helm MergeValues at internal/helm/MergeValues, used
 	//  e.g. in hubble.go. Does using the upstream HelmOpts.MergeValues here create inconsistencies with which
@@ -224,18 +218,9 @@ func (k *K8sInstaller) preinstall(ctx context.Context) error {
 			}
 		}
 	case k8s.KindEKS:
-		chainingMode := getChainingMode(helmValues)
-
-		// Do not stop AWS DS if we are running in chaining mode
-		if chainingMode != "aws-cni" && !k.params.IsDryRun() {
-			if _, err := k.client.GetDaemonSet(ctx, AwsNodeDaemonSetNamespace, AwsNodeDaemonSetName, metav1.GetOptions{}); err == nil {
-				k.Log("🔥 Patching the %q DaemonSet to evict its pods...", AwsNodeDaemonSetName)
-				patch := []byte(fmt.Sprintf(`{"spec":{"template":{"spec":{"nodeSelector":{"%s":"%s"}}}}}`, AwsNodeDaemonSetNodeSelectorKey, AwsNodeDaemonSetNodeSelectorValue))
-				if _, err := k.client.PatchDaemonSet(ctx, AwsNodeDaemonSetNamespace, AwsNodeDaemonSetName, types.StrategicMergePatchType, patch, metav1.PatchOptions{}); err != nil {
-					k.Log("❌ Unable to patch the %q DaemonSet", AwsNodeDaemonSetName)
+		// setup chaining mode
+		if err := k.awsSetupChainingMode(ctx, helmValues); err != nil {
 					return err
-				}
-			}
 		}
 	}
